move pages and page items for reader into subdirectories. rename pages components
