+++ frontend/src/components/reader/pages/horizontal-window.vue
@@ -39,15 +39,15 @@
 import _ from "lodash";
 import { mapActions, mapGetters, mapState } from "pinia";
 
-import BookPage from "@/components/reader/page.vue";
-import PageChangeLink from "@/components/reader/page-change-link.vue";
+import BookPage from "@/components/reader/pages/page/page.vue";
+import PageChangeLink from "@/components/reader/pages/page-change-link.vue";
 import { useReaderStore } from "@/stores/reader";
 
 const WINDOW_BACK_BOUND = 48;
 const WINDOW_FORE_BOUND = 48;
 
 export default {
-  name: "PagesWindowHorizontal",
+  name: "PagesHorizontalWindow",
   components: {
     BookPage,
     PageChangeLink,
