+++ frontend/src/components/reader/pages/page/page-loading.vue
