+++ frontend/src/components/reader/pages/page/page-pdf.vue
